cart option talon (and moving utilities to peregrine)
