+++ packages/peregrine/lib/util/__fixtures__/product.js
