+++ packages/peregrine/lib/util/__tests__/findMatchingProductVariant.spec.js
@@ -1,4 +1,4 @@
-import findMatchingVariant from '../findMatchingProductVariant';
+import { findMatchingVariant } from '../findMatchingProductVariant';
 
 const sampleItem = {
     sku: 'PARENT',
