+++ packages/peregrine/lib/util/__tests__/productVariants.spec.js
