+++ packages/peregrine/lib/util/isProductConfigurable.js
@@ -1,0 +1,3 @@
+// TODO: Move/merge with product util in peregrine?
+export const isProductConfigurable = product =>
+    product.__typename === 'ConfigurableProduct';
