+++ packages/peregrine/lib/util/productVariants.js
