+++ packages/venia-ui/lib/components/MiniCart/cartOptions.js
@@ -1,4 +1,4 @@
-import React, { Suspense, useCallback, useMemo, useState } from 'react';
+import React, { Suspense } from 'react';
 import { array, bool, func, number, shape, string } from 'prop-types';
 import { Form } from 'informed';
 import { Price } from '@magento/peregrine';
@@ -7,11 +7,11 @@ import { mergeClasses } from '../../classify';
 import LoadingIndicator from '../LoadingIndicator';
 import Button from '../Button';
 import Quantity from '../ProductQuantity';
-import appendOptionsToPayload from '../../util/appendOptionsToPayload';
-import { findMatchingProductOptionValue } from '../../util/productVariants';
-import isProductConfigurable from '../../util/isProductConfigurable';
+
+import { isProductConfigurable } from '@magento/peregrine/lib/util/isProductConfigurable';
 
 import defaultClasses from './cartOptions.css';
+import { useCartOptions } from '@magento/peregrine/lib/talons/MiniCart/useCartOptions';
 
 const Options = React.lazy(() => import('../ProductOptions'));
 
@@ -21,22 +21,7 @@ const loadingIndicator = (
     </LoadingIndicator>
 );
 
-const isItemMissingOptions = (cartItem, configItem, numSelections) => {
-    // Non-configurable products can't be missing options
-    if (cartItem.product_type !== 'configurable') {
-        return false;
-    }
-
-    // Configurable products are missing options if we have fewer
-    // option selections than the product has options.
-    const { configurable_options } = configItem;
-    const numProductOptions = configurable_options.length;
-
-    return numSelections < numProductOptions;
-};
-
 const CartOptions = props => {
-    // Props.
     const {
         cartItem,
         configItem,
@@ -45,76 +30,28 @@ const CartOptions = props => {
         isUpdatingItem,
         updateCart
     } = props;
-    const { name, price, qty } = cartItem;
 
-    const initialOptionSelections = useMemo(() => {
-        const result = new Map();
-
-        // This set should contain entries like: 176 => 26, not "Fashion Color" => "Lilac".
-        // To transform, we have to find the matching configurable option and value on the configItem.
-        if (cartItem.options) {
-            cartItem.options.forEach(cartItemOption => {
-                const {
-                    option,
-                    value: optionValue
-                } = findMatchingProductOptionValue({
-                    product: configItem,
-                    variantOption: cartItemOption
-                });
-
-                if (option && optionValue) {
-                    const key = option.attribute_id;
-                    const value = optionValue.value_index;
-                    result.set(key, value);
-                }
-            });
-        }
-
-        return result;
-    }, [cartItem, configItem]);
-
-    // State.
-    const [optionSelections, setOptionSelections] = useState(
-        initialOptionSelections
-    );
-    const [quantity, setQuantity] = useState(qty);
-
-    // Callbacks.
-    const handleSelectionChange = useCallback(
-        (optionId, selection) => {
-            const newSelections = new Map(optionSelections);
-            const lastSelection = Array.from(selection).pop();
-
-            newSelections.set(optionId, lastSelection);
-
-            setOptionSelections(newSelections);
-        },
-        [optionSelections]
-    );
-    const handleUpdateClick = useCallback(() => {
-        const payload = {
-            item: configItem,
-            productType: configItem.__typename,
-            quantity: quantity
-        };
-
-        if (isProductConfigurable(configItem)) {
-            appendOptionsToPayload(payload, optionSelections);
-        }
-
-        updateCart(payload, cartItem.item_id);
-    }, [cartItem, configItem, quantity, optionSelections, updateCart]);
-
-    // Members.
-    const classes = mergeClasses(defaultClasses, props.classes);
-    const isMissingOptions = isItemMissingOptions(
+    const talonProps = useCartOptions({
         cartItem,
         configItem,
-        optionSelections.size
-    );
+        endEditItem,
+        updateCart
+    });
+
+    const {
+        itemName,
+        itemPrice,
+        itemQuantity,
+        handleCancel,
+        handleSelectionChange,
+        handleUpdateClick,
+        handleValueChange,
+        isUpdateDisabled
+    } = talonProps;
+
+    const classes = mergeClasses(defaultClasses, props.classes);
     const modalClass = isUpdatingItem ? classes.modal_active : classes.modal;
 
-    // Render.
     const options = isProductConfigurable(configItem) ? (
         <Suspense fallback={loadingIndicator}>
             <section className={classes.options}>
@@ -130,9 +67,9 @@ const CartOptions = props => {
     return (
         <Form className={classes.root}>
             <div className={classes.focusItem}>
-                <span className={classes.name}>{name}</span>
+                <span className={classes.name}>{itemName}</span>
                 <span className={classes.price}>
-                    <Price currencyCode={currencyCode} value={price} />
+                    <Price currencyCode={currencyCode} value={itemPrice} />
                 </span>
             </div>
             <div className={classes.form}>
@@ -141,17 +78,20 @@ const CartOptions = props => {
                     <h2 className={classes.quantityTitle}>
                         <span>Quantity</span>
                     </h2>
-                    <Quantity initialValue={qty} onValueChange={setQuantity} />
+                    <Quantity
+                        initialValue={itemQuantity}
+                        onValueChange={handleValueChange}
+                    />
                 </section>
             </div>
             <div className={classes.save}>
-                <Button onClick={endEditItem}>
+                <Button onClick={handleCancel}>
                     <span>Cancel</span>
                 </Button>
                 <Button
                     priority="high"
                     onClick={handleUpdateClick}
-                    disabled={isMissingOptions}
+                    disabled={isUpdateDisabled}
                 >
                     <span>Update Cart</span>
                 </Button>
