--- packages/venia-ui/lib/util/isProductConfigurable.js
@@ -1,4 +1,0 @@
-const isProductConfigurable = product =>
-    product.__typename === 'ConfigurableProduct';
-
-export default isProductConfigurable;
